# Add faster c implementations for some functions in triangulation.py