I implemented some of the functions in triangulation.py in C for a considerable speed boost in several situations. They may not support all list types that are used, but I tried to implement functionality for the ones I saw.

The timing differences are as follows, for 100000 iterations of arbitrarily chosen function arguments:
```
         C                         fast_norm:     0.025422s
Python                         fast_norm:     0.074730s
         C         fast_2d_circumcircle:     0.035352s
Python         fast_2d_circumcircle:     1.077767s
         C         fast_3d_circumcircle:     0.068189s
Python         fast_3d_circumcircle:     1.729581s
         C  fast_2d_point_in_simplex:     0.037834s
Python  fast_2d_point_in_simplex:     0.170977s
```

The difference is considerably large for fast_2d_circumcircle and fast_3d_circumcircle, as these use numpy array operations that massively slow the program down - getting rid of

```
    points = np.array(points)
    pts = points[1:] - points[0]
```

in both of these functions would provide a considerable speedup - converting points to an array creates a lot of unnecessary overhead.