@basnijholt 2019-12-15
This is awesome! I'll check it out in more detail when I'm near my computer.